Chapter 11 Examples to Start
Showing off some group boxes and separators. The vertical separator shows; the horizontal, eh, not so much. Could be my placement methodology, could be some quirky thing with tkinter. I've already noticed that the group boxes need contents to show and do not appear unconditionally, which seems contrary to what the book shows. This was built on Windows 10, so maybe a different outcome on Linux?